search with dto and paged  result on Author

+++ src/main/java/com/alibou/security/section/SectionRepository.java
@@ -2,5 +2,5 @@ package com.alibou.security.section;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectionRepository extends JpaRepository<Section, Long> {
+public interface SectionRepository extends JpaRepository<Section, Integer> {
 }
